Rethink the "clear" functionality
The "clear" method was behaving more like a "destroy" without a subsequent
re-initialize.  This change keeps clear working by simply invoking the clear
method on the vcs side, then rendering the cleared result to the canvas.  As
a result, it also fixes the issue that nothing could be drawn after invoking
clear.