Fix magma colormap issue
The first time the "mycolormap" link was clicked, it worked fine, but
every time after that, it did nothing but generated a stacktrace in the
console.  This was because "mycolormap" had already been created.

This change allows you to click "mycolormap" more than once.  For example,
if you have clicked the "Clear" link, you may want to click the "mycolormap"
link a second time.